worker/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-elevenlabs
livekit-plugins-silero
python-dotenv
uvicorn[standard]